cli/client/requirements.txt
pyyaml
typer
shellingham
rich
emoji
dictor
